Rename VMI reconciler
It reconciles VMI not VM, hence rename accordingly.
Rename folder, files and package accordingly.
